+++ sdks/python/apache_beam/runners/common_test.py
@@ -19,26 +19,16 @@
 
 from __future__ import absolute_import
 
-import threading
-import time
 import unittest
 
 import apache_beam as beam
-from apache_beam.io.concat_source_test import RangeSource
-from apache_beam.io.restriction_trackers import OffsetRange
-from apache_beam.io.restriction_trackers import OffsetRestrictionTracker
-from apache_beam.io.watermark_estimators import ManualWatermarkEstimator
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.runners.common import DoFnSignature
-from apache_beam.runners.common import _RestrictionTrackerView
-from apache_beam.runners.common import _ThreadsafeRestrictionTracker
-from apache_beam.runners.common import _ThreadsafeWatermarkEstimator
 from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.test_stream import TestStream
 from apache_beam.transforms import trigger
 from apache_beam.transforms import window
 from apache_beam.transforms.core import DoFn
-from apache_beam.utils import timestamp
 
 
 class DoFnSignatureTest(unittest.TestCase):
@@ -120,114 +110,6 @@ class DoFnProcessTest(unittest.TestCase):
       test_stream = (TestStream().advance_watermark_to(10).add_elements([1, 2]))
       (p | test_stream | beam.ParDo(DoFnProcessWithKeyparam()))
 
-
-class _ThreadsafeRestrictionTrackerTest(unittest.TestCase):
-
-  def test_initialization(self):
-    with self.assertRaises(ValueError):
-      _ThreadsafeRestrictionTracker(RangeSource(0, 1), threading.Lock())
-
-  def test_defer_remainder_with_wrong_time_type(self):
-    threadsafe_tracker = _ThreadsafeRestrictionTracker(
-        OffsetRestrictionTracker(OffsetRange(0, 10)), threading.Lock())
-    with self.assertRaises(ValueError):
-      threadsafe_tracker.defer_remainder(10)
-
-  def test_self_checkpoint_immediately(self):
-    restriction_tracker = OffsetRestrictionTracker(OffsetRange(0, 10))
-    threadsafe_tracker = _ThreadsafeRestrictionTracker(
-        restriction_tracker, threading.Lock())
-    threadsafe_tracker.defer_remainder()
-    deferred_residual, deferred_time = threadsafe_tracker.deferred_status()
-    expected_residual = OffsetRange(0, 10)
-    self.assertEqual(deferred_residual, expected_residual)
-    self.assertTrue(isinstance(deferred_time, timestamp.Duration))
-    self.assertEqual(deferred_time, 0)
-
-  def test_self_checkpoint_with_relative_time(self):
-    threadsafe_tracker = _ThreadsafeRestrictionTracker(
-        OffsetRestrictionTracker(OffsetRange(0, 10)), threading.Lock())
-    threadsafe_tracker.defer_remainder(timestamp.Duration(100))
-    time.sleep(2)
-    _, deferred_time = threadsafe_tracker.deferred_status()
-    self.assertTrue(isinstance(deferred_time, timestamp.Duration))
-    # The expectation = 100 - 2 - some_delta
-    self.assertTrue(deferred_time <= 98)
-
-  def test_self_checkpoint_with_absolute_time(self):
-    threadsafe_tracker = _ThreadsafeRestrictionTracker(
-        OffsetRestrictionTracker(OffsetRange(0, 10)), threading.Lock())
-    now = timestamp.Timestamp.now()
-    schedule_time = now + timestamp.Duration(100)
-    self.assertTrue(isinstance(schedule_time, timestamp.Timestamp))
-    threadsafe_tracker.defer_remainder(schedule_time)
-    time.sleep(2)
-    _, deferred_time = threadsafe_tracker.deferred_status()
-    self.assertTrue(isinstance(deferred_time, timestamp.Duration))
-    # The expectation =
-    # schedule_time - the time when deferred_status is called - some_delta
-    self.assertTrue(deferred_time <= 98)
-
-
-class RestrictionTrackerViewTest(unittest.TestCase):
-
-  def test_initialization(self):
-    with self.assertRaises(ValueError):
-      _RestrictionTrackerView(OffsetRestrictionTracker(OffsetRange(0, 10)))
-
-  def test_api_expose(self):
-    threadsafe_tracker = _ThreadsafeRestrictionTracker(
-        OffsetRestrictionTracker(OffsetRange(0, 10)), threading.Lock())
-    tracker_view = _RestrictionTrackerView(threadsafe_tracker)
-    current_restriction = tracker_view.current_restriction()
-    self.assertEqual(current_restriction, OffsetRange(0, 10))
-    self.assertTrue(tracker_view.try_claim(0))
-    tracker_view.defer_remainder()
-    deferred_remainder, deferred_watermark = (
-        threadsafe_tracker.deferred_status())
-    self.assertEqual(deferred_remainder, OffsetRange(1, 10))
-    self.assertEqual(deferred_watermark, timestamp.Duration())
-
-  def test_non_expose_apis(self):
-    threadsafe_tracker = _ThreadsafeRestrictionTracker(
-        OffsetRestrictionTracker(OffsetRange(0, 10)), threading.Lock())
-    tracker_view = _RestrictionTrackerView(threadsafe_tracker)
-    with self.assertRaises(AttributeError):
-      tracker_view.check_done()
-    with self.assertRaises(AttributeError):
-      tracker_view.current_progress()
-    with self.assertRaises(AttributeError):
-      tracker_view.try_split()
-    with self.assertRaises(AttributeError):
-      tracker_view.deferred_status()
-
-
-class ThreadsafeWatermarkEstimatorTest(unittest.TestCase):
-
-  def test_initialization(self):
-    with self.assertRaises(ValueError):
-      _ThreadsafeWatermarkEstimator(None, threading.Lock())
-
-  def test_get_estimator_state(self):
-    estimator = _ThreadsafeWatermarkEstimator(
-        ManualWatermarkEstimator(None), threading.Lock())
-    self.assertIsNone(estimator.get_estimator_state())
-    estimator.set_watermark(timestamp.Timestamp(10))
-    self.assertEqual(estimator.get_estimator_state(), timestamp.Timestamp(10))
-
-  def test_track_timestamp(self):
-    estimator = _ThreadsafeWatermarkEstimator(
-        ManualWatermarkEstimator(None), threading.Lock())
-    estimator.observe_timestamp(timestamp.Timestamp(10))
-    self.assertIsNone(estimator.current_watermark())
-    estimator.set_watermark(timestamp.Timestamp(20))
-    self.assertEqual(estimator.current_watermark(), timestamp.Timestamp(20))
-
-  def test_non_exsited_attr(self):
-    estimator = _ThreadsafeWatermarkEstimator(
-        ManualWatermarkEstimator(None), threading.Lock())
-    with self.assertRaises(AttributeError):
-      estimator.non_existed_call()
 
 if __name__ == '__main__':
   unittest.main()
